deleted unnecessary files that were created on default, moved test files into a 'misc' folder, and deleted css files (since we're using styled-components)
