+++ frontend/src/misc/reportWebVitals.js
